Update to add support for automated builds via Travis-CI.
As part of this effort, we've needed to nail down the versions of the
libraries we linked with, and do a lot of back-and-forth about
supported versions. In the end we only support 7.10 and 8.0 because
7.8 appears to trigger an error during macro processing on Options.hs.
(It doesn't do this on my computer, but repeatably does it on the
Travis build machines.) Also, for fun, we tweaked some of the
formatting of braces on data definitions.

(This is a squash commit of the 'travis' branch.)