Add automatic packaging support, to make distribution easier, and add a note in the README.
This update adds automatic RPM and Debian/Ubuntu packaging, so that we can
automate the release process a bit. The main remaining issue is that many of the
systems we might want to target include GHC versions too old to build this. So
we might want to investigate how to build something for those, possibly by
building our own versions of GHC and friends?

As a side note, this includes a manual set of Lens derivations because of
problems found when compiling with GHC 7.8.4.

(This is a squash merge of the 'packaging' branch.)